services/ingestion/requirements.txt
fastapi>=0.135.0
uvicorn>=0.41.0
python-multipart>=0.0.22
pydantic>=2.12.0
pydantic-settings>=2.13.0
PyPDF2>=3.0.1
python-docx>=1.2.0
beautifulsoup4>=4.14.0
requests>=2.32.0
httpx>=0.28.0
aiofiles>=25.1.0
python-dotenv>=1.2.0
deepgram-sdk>=6.0.0
yt-dlp>=2024.0.0
